Revert "[Refactor] centralize extensions"
