--- .flaskenv
@@ -1,3 +1,0 @@
-FLASK_APP=run.py
-FLASK_RUN_PORT=8080
-
