--- .vscode/launch.json
@@ -1,21 +1,0 @@
-{
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "Flask run",
-            "type": "python",
-            "request": "launch",
-            "module": "flask",
-            "envFile": "${workspaceFolder}/.env",
-            "env": {
-                "FLASK_APP": "run.py",
-                "FLASK_ENV": "development",
-                "FLASK_RUN_PORT": "8080"
-            },
-            "args": [
-                "run"
-            ],
-            "jinja": true
-        }
-    ]
-}
