+++ crunevo/models/__init__.py
@@ -1,1 +1,4 @@
-from crunevo.extensions import db
+from flask_sqlalchemy import SQLAlchemy
+
+
+db = SQLAlchemy()
