+++ crunevo/models/post.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from crunevo.extensions import db
+from crunevo.models import db
 
 
 class Post(db.Model):
