+++ crunevo/models/post_comment.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from crunevo.extensions import db
+from crunevo.models import db
 
 
 class PostComment(db.Model):
