+++ crunevo/tests/test_app.py
@@ -2,7 +2,7 @@ import os
 import pytest
 
 from crunevo.app import create_app
-from crunevo.extensions import db
+from crunevo.models import db
 
 
 @pytest.fixture
