+++ crunevo/tests/test_post_comments.py
@@ -1,7 +1,7 @@
 import os
 import pytest
 from crunevo.app import create_app
-from crunevo.extensions import db
+from crunevo.models import db
 from crunevo.models.user import User
 from crunevo.models.post import Post
 from crunevo.models.post_comment import PostComment
