+++ crunevo/tests/test_upload.py
@@ -3,7 +3,7 @@ import io
 import pytest
 
 from crunevo.app import create_app
-from crunevo.extensions import db
+from crunevo.models import db
 from crunevo.models.user import User
